# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cord
    description: CORD AI agent operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /cord/ai:
    post:
      operationId: cordAiQuery
      tags: [cord]
      summary: Query CORD AI agent
      description: Sends transcript to Gemini and returns the AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CordAiRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CordAiResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cord/tts:
    post:
      operationId: cordTts
      tags: [cord]
      summary: Convert text to speech
      description: Sends text to ElevenLabs TTS and returns audio data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CordTtsRequest"
      responses:
        "200":
          description: Audio data as base64
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CordTtsResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CordAiRequest:
      type: object
      properties:
        transcript:
          type: string
          description: The user's spoken transcript
        geminiApiKey:
          type: string
          description: Optional Gemini API key override (for demo purposes)
      required:
        - transcript
    CordAiResponse:
      type: object
      properties:
        response:
          type: string
          description: The AI agent's response text
      required:
        - response
    CordTtsRequest:
      type: object
      properties:
        text:
          type: string
          description: Text to convert to speech
        elevenLabsApiKey:
          type: string
          description: Optional ElevenLabs API key override (for demo purposes)
      required:
        - text
    CordTtsResponse:
      type: object
      properties:
        audioBase64:
          type: string
          description: Base64-encoded audio data (MP3)
        contentType:
          type: string
          description: Audio content type
      required:
        - audioBase64
        - contentType
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error