# lib/api-zod/src/generated/types/cordAiRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CordAiRequest {
  /** The user's spoken transcript */
  transcript: string;
  /** Optional Gemini API key override (for demo purposes) */
  geminiApiKey?: string;
}