# lib/api-zod/src/generated/types/cordAiResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CordAiResponse {
  /** The AI agent's response text */
  response: string;
}