# lib/api-zod/src/generated/types/cordTtsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CordTtsRequest {
  /** Text to convert to speech */
  text: string;
  /** Optional ElevenLabs API key override (for demo purposes) */
  elevenLabsApiKey?: string;
}