# lib/api-zod/src/generated/types/cordTtsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CordTtsResponse {
  /** Base64-encoded audio data (MP3) */
  audioBase64: string;
  /** Audio content type */
  contentType: string;
}